Move is_mingw_gnu_toolchain and i686_decorated_name to cg_ssa
